requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.17.1
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
psycopg2==2.9.11
Pygments==2.20.0
python-decouple==3.8
sqlparse==0.5.5
